dasshio/requirements.txt
netifaces
requests
scapy-python3
